% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{CFPCA}
\alias{CFPCA}
\title{Componentwise Multivariate Functional Principal Component Analysis}
\usage{
CFPCA(Ly, Lt, optns = list())
}
\arguments{
\item{Ly}{A list of matrices, each being D by n_i containing the observed values for individual i. In each matrix, columes corresponds to different time points, and rows corresponds to different dimensions.}

\item{Lt}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y. Each vector should be sorted in ascending order.}

\item{optns}{A list of options control parameters specified by \code{list(name=value)}. See the `Details' section in `?FPCA`.}
}
\value{
See `Return` in `?RFPCA`
}
\description{
Implements the method of Happs and Grevens (2018) for Euclidean-valued multivariate functional data.
}
