% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{coordToTanV}
\alias{coordToTanV}
\alias{coordToTanV.default}
\title{Transform the coordinates for the tangent space into a tangent vector}
\usage{
coordToTanV(mfd, p, V0, ...)

\method{coordToTanV}{default}(mfd, p, V0)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{p}{A vector for the base point}

\item{V0}{A matrix or a vector containing the coordinates}
}
\description{
Transform the coordinates for the tangent space into a tangent vector
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Method
}}

