% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuManifold.R, R/General.R, R/SOManifold.R,
%   R/SPDManifolds.R, R/SphereManifold.R
\name{distance.Euclidean}
\alias{distance.Euclidean}
\alias{distance}
\alias{distance.SO}
\alias{distance.LogEu}
\alias{distance.AffInv}
\alias{distance.Sphere}
\title{Geodesic distance}
\usage{
\method{distance}{Euclidean}(mfd, X, Y)

distance(mfd, X, Y)

\method{distance}{SO}(mfd, X, Y)

\method{distance}{LogEu}(mfd, X, Y, assumeLogRep = FALSE)

\method{distance}{AffInv}(mfd, X, Y)

\method{distance}{Sphere}(mfd, X, Y)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{X, Y}{Matrices with n columns. The distance between each pair of columns is calculated. If either X or Y is a vector then it is recycled.}
}
\description{
Geodesic distance
}
\section{Methods (by class)}{
\itemize{
\item \code{Euclidean}: Method

\item \code{SO}: Method

\item \code{LogEu}: Method

\item \code{AffInv}: Method

\item \code{Sphere}: Method
}}

