% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitted.R
\name{fitted.RFPCA}
\alias{fitted.RFPCA}
\title{Get the fitted trajectories for an RFPCA object}
\usage{
\method{fitted}{RFPCA}(object, K, grid = c("obs", "work"), ...)
}
\arguments{
\item{K}{The number of components to use}

\item{grid}{Either `obs` for obsGrid or `work` for workGrid}
}
\description{
Get the fitted trajectories for an RFPCA object
}
