% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General.R
\name{frechetMeanCurve}
\alias{frechetMeanCurve}
\title{Frechet mean curve of a set of sparsely observed curves on the manifold}
\usage{
frechetMeanCurve(
  mfd,
  bw,
  kernel_type,
  xin,
  yin,
  xout,
  npoly = 1L,
  win = rep(1L, length(xin))
)
}
\arguments{
\item{mfd}{A class instance that represents the Riemannian manifold. See the {manifold} package}

\item{bw}{The bandwidth}

\item{kernel_type}{The type of kernel for smoothing}

\item{xin}{A list of \emph{n} vectors containing the observation time points for each individual corresponding to y. Each vector should be sorted in ascending order.}

\item{yin}{A list of \emph{n} matrices containing the observed values for each individual. Missing values specified by \code{NA}s are supported for dense case (\code{dataType='dense'}).}

\item{xout}{The time points where the mean curve shall be evaluated}

\item{npoly}{The order of the local polynomial smoother}

\item{win}{A vector \emph{n} numbers. The weight of each individual.}
}
\value{
A matrix with \code{length(xout)} columns. The \emph{i}th column is the estimated mean curve at \code{xout[i]}
}
\description{
Frechet mean curve of a set of sparsely observed curves on the manifold
}
