% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereManifold.R
\name{frechetMeanCurve.Sphere}
\alias{frechetMeanCurve.Sphere}
\title{Frechet mean of a set of points}
\usage{
frechetMeanCurve.Sphere(mfd, bw, kernel_type, xin, yin, xout, npoly = 1L,
  win = rep(1L, length(xin)))
}
\description{
Frechet mean of a set of points
}
