% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogM}
\alias{LogM}
\title{Matrix logarithm}
\usage{
LogM(X)
}
\arguments{
\item{X}{A square matrix}
}
\description{
\code{LogM(X)} computes the matrix logarithm of a general matrix using Armardillo. The returned value is a complex matrix.
}
