% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogMSPD}
\alias{LogMSPD}
\title{Matrix logarithm of a symmetric positive definite}
\usage{
LogMSPD(X)
}
\description{
\code{LogMSPD(X)} computes the matrix logarithm of an SPD matrix using Armardillo. The returned value is a real matrix.
}
