% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeInput.R
\name{MakeInput}
\alias{MakeInput}
\title{Create RFPCA input from a data frame}
\usage{
MakeInput(data, IDvar, tvar, yvar)
}
\arguments{
\item{data}{A data frame}

\item{IDvar}{The column name that stands for the ID}

\item{tvar}{The column name that stands for the time}

\item{yvar}{The column names that stands for the values for Lt. If not specified, all columns except for IDvar and tvar are used.}
}
\description{
Create RFPCA input from a data frame
}
