% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{Makemu}
\alias{Makemu}
\title{Create a mean function}
\usage{
Makemu(mfd, VtList, p0, pts = seq(0, 1, length.out = 50))
}
\arguments{
\item{mfd}{An object whose class specifies the manifold to use}

\item{VtList}{A list of functions. The jth function takes a vector of time, and return the corresponding V(t) in the jth entry}

\item{p0}{The basepoint, which will be converted to a vector}

\item{pts}{The time points to generate the mean function}
}
\description{
Create a mean function
}
