% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuManifold.R, R/General.R, R/SOManifold.R,
%   R/SPDManifolds.R, R/SphereManifold.R
\name{metric.Euclidean}
\alias{metric.Euclidean}
\alias{metric}
\alias{metric.default}
\alias{metric.SO}
\alias{metric.LogEu}
\alias{metric.AffInv}
\alias{metric.Sphere}
\title{Returns the Riemannian metric}
\usage{
\method{metric}{Euclidean}(mfd, p, U, V)

metric(mfd, p, U, V)

\method{metric}{default}(mfd, p = NULL, U, V)

\method{metric}{SO}(mfd, p, U, V)

\method{metric}{LogEu}(mfd, p, U, V)

\method{metric}{AffInv}(mfd, p, U, V)

\method{metric}{Sphere}(mfd, p, U, V)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{p}{A vector containing the (single) base point on the manifold.}

\item{U, V}{Matrices with n columns each containing a tangent vector. The metric of each pair of n columns is calculated.}
}
\description{
Returns the Riemannian metric
}
\section{Methods (by class)}{
\itemize{
\item \code{Euclidean}: Method

\item \code{default}: Method

\item \code{SO}: Method

\item \code{LogEu}: Method

\item \code{AffInv}: Method

\item \code{Sphere}: Method
}}

