% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuManifold.R, R/General.R, R/SPDManifolds.R,
%   R/SphereManifold.R
\name{NoiseTangent.Euclidean}
\alias{NoiseTangent.Euclidean}
\alias{NoiseTangent}
\alias{NoiseTangent.SPD}
\alias{NoiseTangent.Sphere}
\title{Generate noise on the tangent space}
\usage{
\method{NoiseTangent}{Euclidean}(mfd, n, mu, sigma2 = 0, epsFun = rnorm)

NoiseTangent(mfd, n, mu, sigma2 = 0, epsFun = rnorm)

\method{NoiseTangent}{SPD}(mfd, n, mu, sigma2 = 0, epsFun = rnorm)

\method{NoiseTangent}{Sphere}(mfd, n, mu, sigma2 = 0, epsFun = rnorm)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{n}{Sample size}

\item{mu}{A matrix or a vector containing the base point(s)}

\item{epsFun}{The distribution for generating the univariate errors with unit variance for the tangent space coordinates}

\item{sigam2}{The isotropic variance on the tangent space}
}
\value{
An array with size (n, dimIntrinsic, nPts)
}
\description{
For each time point (column in mu), generate n isotropic noises distributed as epsFun on the tangent space at mu(t)
}
\section{Methods (by class)}{
\itemize{
\item \code{Euclidean}: Method

\item \code{SPD}: Method

\item \code{Sphere}: Method
}}

