% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuManifold.R, R/General.R, R/SOManifold.R,
%   R/SPDManifolds.R, R/SphereManifold.R
\name{origin.Euclidean}
\alias{origin.Euclidean}
\alias{origin}
\alias{origin.SO}
\alias{origin.SPD}
\alias{origin.Sphere}
\title{Returns the origin of the manifold}
\usage{
\method{origin}{Euclidean}(mfd, dimIntrinsic)

origin(mfd, dimIntrinsic, ...)

\method{origin}{SO}(mfd, dimIntrinsic)

\method{origin}{SPD}(mfd, dimIntrinsic)

\method{origin}{Sphere}(mfd, dimIntrinsic)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}
}
\description{
Each manifold defined in this package has been designated a more or less reasonable choice of origin.
}
\section{Methods (by class)}{
\itemize{
\item \code{Euclidean}: Method

\item \code{SO}: The origin has 1 in the first ambient coordinate and 0 otherwise.

\item \code{SPD}: The origin is the identity matrix.

\item \code{Sphere}: The origin has 1 in the first ambient coordinate and 0 otherwise.
}}

