% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.RFPCA.R
\name{plot.RFPCA}
\alias{plot.RFPCA}
\title{Plot an RFPCA object}
\usage{
\method{plot}{RFPCA}(
  x,
  type = c("mu", "phi", "cov", "fitted", "raw"),
  K = 3,
  subset,
  inputData,
  dimNames,
  ...
)
}
\arguments{
\item{x}{An RFPCA object}

\item{type}{The type of plot to produce. 'mu', mean function; 'phi', eigenfunctions; 'cov', covariance function'; 'fitted', fitted trajectories; 'raw', raw observations}

\item{K}{The number of components to use or plot}

\item{subset}{A subset of subject IDs to plot}

\item{inputData}{The input data (a list of Ly and Lt)}

\item{dimNames}{An optional vector specifying the dimension names}

\item{...}{Not used}
}
\description{
Plot an RFPCA object
}
