% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict.RFPCA}
\alias{predict.RFPCA}
\title{Obtains the predicted value for a set of new sparse observations}
\usage{
\method{predict}{RFPCA}(
  object,
  newLy,
  newLt,
  sigma2 = NULL,
  K,
  xiMethod = c("CE", "IN"),
  type = c("xi", "traj"),
  ...
)
}
\arguments{
\item{object}{An `RFPCA` object}

\item{newLy}{A list containing the new response values; each entry corresponds to a subject and should be a matrix in which each column is an observation.}

\item{newLt}{A list containing the new time points; each entry corresponds to a subject and should be a vector.}

\item{sigma2}{The noise variance to use for the prediction. Serves as a regularization parameter. If left as `NULL`, use the one chosen when fitting `object`.}

\item{K}{The number of components to apply. If missing, default to all components in `object`}

\item{xiMethod}{Either to use the conditional expectation (`'CE'`) or the integration method (`'IN'`) to predict}

\item{type}{Either to predict the FPC scores (`'xi'`) or the underlying trajectories (`'traj'`)}
}
\value{
A 2D array of predicted scores with entries corresponds to (i, k)
}
\description{
Obtains the predicted value for a set of new sparse observations
}
