% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuManifold.R, R/General.R, R/SOManifold.R,
%   R/SPDManifolds.R, R/SphereManifold.R
\name{project.Euclidean}
\alias{project.Euclidean}
\alias{project}
\alias{project.SO}
\alias{project.AffInv}
\alias{project.LogEu}
\alias{project.SPD}
\alias{project.Sphere}
\title{Project data points in the ambient space onto the manifold}
\usage{
\method{project}{Euclidean}(mfd, A)

project(mfd, p)

\method{project}{SO}(mfd, A)

\method{project}{AffInv}(mfd, A)

\method{project}{LogEu}(mfd, A)

\method{project}{SPD}(mfd, A)

\method{project}{Sphere}(mfd, p)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{p}{A matrix of points to be projected to the manifold.}
}
\description{
Project data points in the ambient space onto the manifold
}
\section{Methods (by class)}{
\itemize{
\item \code{Euclidean}: Method

\item \code{SO}: Method

\item \code{AffInv}: Method

\item \code{LogEu}: Method

\item \code{SPD}: Method

\item \code{Sphere}: Method
}}

