% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuManifold.R, R/General.R, R/SOManifold.R,
%   R/SPDManifolds.R, R/SphereManifold.R
\name{projectTangent.Euclidean}
\alias{projectTangent.Euclidean}
\alias{projectTangent}
\alias{projectTangent.SO}
\alias{projectTangent.SPD}
\alias{projectTangent.Sphere}
\title{Project data points in the ambient space onto the tangent space}
\usage{
\method{projectTangent}{Euclidean}(mfd, p, X, projMatOnly = FALSE)

projectTangent(mfd, ...)

\method{projectTangent}{SO}(mfd, p, X, projMatOnly = FALSE)

\method{projectTangent}{SPD}(mfd, p, X, projMatOnly = FALSE)

\method{projectTangent}{Sphere}(mfd, p, X, projMatOnly = FALSE)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{p}{A vector containing the base point on the manifold. Data X will be projected onto the tangent space of p.}

\item{X}{A vector or a matrix containing the data points in terms of the coordinates in the ambient space.}

\item{projMatOnly}{Whether to only return the projection matrix (FALSE) or the projected data points (TRUE, the default)}
}
\description{
Project data points in the ambient space onto the tangent space
}
\section{Methods (by class)}{
\itemize{
\item \code{Euclidean}: Method

\item \code{SO}: Method

\item \code{SPD}: Method

\item \code{Sphere}: Method
}}

