% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuManifold.R, R/General.R, R/SOManifold.R,
%   R/SPDManifolds.R, R/SphereManifold.R
\name{rieExp.Euclidean}
\alias{rieExp.Euclidean}
\alias{rieExp}
\alias{rieExp.SO}
\alias{rieExp.LogEu}
\alias{rieExp.AffInv}
\alias{rieExp.Sphere}
\title{Riemannian exponential map}
\usage{
\method{rieExp}{Euclidean}(mfd, p, V)

rieExp(mfd, p, V)

\method{rieExp}{SO}(mfd, p, V, tol = 1e-10)

\method{rieExp}{LogEu}(mfd, p, V, tol = 1e-10)

\method{rieExp}{AffInv}(mfd, p, V, tol = 1e-10)

\method{rieExp}{Sphere}(mfd, p, V)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{p, V}{Matrices with n columns. The exponential map of each column in V is calculated w.r.t. the corresponding column in p.}
}
\description{
Riemannian exponential map
}
\section{Methods (by class)}{
\itemize{
\item \code{Euclidean}: Method

\item \code{SO}: Method

\item \code{LogEu}: Method

\item \code{AffInv}: Method

\item \code{Sphere}: Method
}}

