% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EuManifold.R, R/General.R, R/SOManifold.R,
%   R/SPDManifolds.R, R/SphereManifold.R
\name{rieLog.Euclidean}
\alias{rieLog.Euclidean}
\alias{rieLog}
\alias{rieLog.SO}
\alias{rieLog.LogEu}
\alias{rieLog.AffInv}
\alias{rieLog.Sphere}
\title{Riemannian logarithm map}
\usage{
\method{rieLog}{Euclidean}(mfd, p, X)

rieLog(mfd, p, X)

\method{rieLog}{SO}(mfd, p, X, tol = 1e-10)

\method{rieLog}{LogEu}(mfd, p, X, tol = 1e-10)

\method{rieLog}{AffInv}(mfd, p, X, tol = 1e-10)

\method{rieLog}{Sphere}(mfd, p, X, tol = 1e-10)
}
\arguments{
\item{mfd}{A manifold object created by \code{\link{createM}}}

\item{p, X}{Matrices with n columns. The logarithm map of each column in X is calculated w.r.t. the corresponding column in p.}
}
\description{
Riemannian logarithm map
}
\section{Methods (by class)}{
\itemize{
\item \code{Euclidean}: Method

\item \code{SO}: Method

\item \code{LogEu}: Method

\item \code{AffInv}: Method

\item \code{Sphere}: Method
}}

