% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphereManifold.R
\name{runifSphere}
\alias{runifSphere}
\title{Generate uniform random variables on the unit sphere}
\usage{
runifSphere(n, dimAmbient)
}
\arguments{
\item{n}{Sample size}

\item{dimAmbient}{The dimension of the ambient space}
}
\description{
Generate uniform random variables on the unit sphere
}
