% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitted.R
\name{fitted.RFPCA}
\alias{fitted.RFPCA}
\title{Obtains the fitted values from an `RFPCA` object}
\usage{
\method{fitted}{RFPCA}(object, K, grid = c("work", "obs"), ...)
}
\arguments{
\item{object}{An `RFPCA` object}

\item{K}{The number of components to apply. If missing, default to all components in `object`}

\item{grid}{The grid where the output function should be supported on}

\item{...}{Not used}
}
\value{
A 3D array with entries corresponds to (i, j, t) supported on workGrid
}
\description{
Obtains the fitted values from an `RFPCA` object
}
